+++ CHANGELOG.md
@@ -21,6 +21,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Changed
 
 * Improving OpenAPI specification
+* Reimplementing database seeding using standalone class and expanded, generic, predictable resources 
 * Importing Grant (Project) type enum from Grant provider, rather than duplicating in Project provider